flux-jobs: return non-zero when no jobID found
Problem: flux jobs JOBID exits with success even when JOBID cannot be
found. It might be nice for the case where if a single JOBID or ALL
JOBIDs cannot be found to return a non-zero exit code.

Add a check in flux-jobs.py to see if all of the JOBIDs passed cannot be
found, and if so, return non-zero. Adjust the tests in t2800-flux-jobs.t
to account for the change in behavior; in particular, the tests that
pass an unknown JOBID and expect success.